JBTM-2538 Isolate product tests from each other
