+++ narayana/ArjunaJTA/jta/tests/classes/io/narayana/perf/product/BitronixComparison.java
@@ -1,0 +1,98 @@
+/*
+ * JBoss, Home of Professional Open Source.
+ * Copyright 2014, Red Hat, Inc., and individual contributors
+ * as indicated by the @author tags. See the copyright.txt file in the
+ * distribution for a full listing of individual contributors.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+ */
+package io.narayana.perf.product;
+
+import bitronix.tm.BitronixTransactionManager;
+import bitronix.tm.TransactionManagerServices;
+import bitronix.tm.utils.DefaultExceptionAnalyzer;
+import com.arjuna.ats.arjuna.common.CoreEnvironmentBeanException;
+import com.arjuna.ats.jta.xa.performance.JMHConfigJTA;
+import com.arjuna.ats.jta.xa.performance.XAResourceImpl;
+import org.junit.Test;
+import org.openjdk.jmh.annotations.Benchmark;
+import org.openjdk.jmh.annotations.Scope;
+import org.openjdk.jmh.annotations.State;
+import org.openjdk.jmh.runner.RunnerException;
+import org.openjdk.jmh.runner.options.CommandLineOptionException;
+
+import javax.transaction.TransactionManager;
+import javax.transaction.UserTransaction;
+import javax.transaction.xa.XAResource;
+
+@State(Scope.Benchmark)
+public class BitronixComparison extends ProductComparison {
+    public static void main(String[] args) throws RunnerException, CommandLineOptionException, CoreEnvironmentBeanException {
+        JMHConfigJTA.runJTABenchmark(BitronixComparison.class.getSimpleName(), args);
+    }
+
+    protected ProductInterface getProductInterface() {
+        return bitronix;
+    }
+
+    protected ProductWorker<Void> getProductWorker() {
+        return new BitronixWorker<Void>(getProductInterface());
+    }
+
+    private ProductInterface bitronix = new ProductInterface() {
+        private BitronixTransactionManager btm;
+
+        @Override
+        public UserTransaction getUserTransaction() {
+            return btm;
+        }
+
+        @Override
+        public TransactionManager getTransactionManager() {
+            return btm;
+        }
+
+        @Override
+        public String getName() {
+            return "bitronix";
+        }
+
+        @Override
+        public String getNameOfMetric() {
+            return outerClassName + "_Bitronix";
+        }
+
+        @Override
+        public XAResource getXAResource() {
+            return new XAResourceImpl();
+        }
+
+        @Override
+        public void init() {
+            TransactionManagerServices.getConfiguration().setGracefulShutdownInterval(1);
+            TransactionManagerServices.getConfiguration().setExceptionAnalyzer(DefaultExceptionAnalyzer.class.getName());
+            btm = TransactionManagerServices.getTransactionManager();
+        }
+
+        @Override
+        public void fini() {
+            if (btm != null) {
+                btm.shutdown();
+                btm = null;
+            }
+        }
+    };
+}
