+++ narayana/ArjunaJTA/jta/tests/classes/io/narayana/perf/product/JotmComparison.java
@@ -1,0 +1,100 @@
+/*
+ * JBoss, Home of Professional Open Source.
+ * Copyright 2014, Red Hat, Inc., and individual contributors
+ * as indicated by the @author tags. See the copyright.txt file in the
+ * distribution for a full listing of individual contributors.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+ */
+package io.narayana.perf.product;
+
+import com.arjuna.ats.arjuna.common.CoreEnvironmentBeanException;
+
+import com.arjuna.ats.jta.xa.performance.JMHConfigJTA;
+import com.arjuna.ats.jta.xa.performance.XAResourceImpl;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.openjdk.jmh.annotations.Benchmark;
+import org.openjdk.jmh.annotations.Scope;
+import org.openjdk.jmh.annotations.Setup;
+import org.openjdk.jmh.annotations.State;
+
+import org.openjdk.jmh.runner.RunnerException;
+import org.openjdk.jmh.runner.options.CommandLineOptionException;
+
+import javax.transaction.SystemException;
+import javax.transaction.TransactionManager;
+import javax.transaction.UserTransaction;
+import javax.transaction.xa.XAResource;
+
+@State(Scope.Benchmark)
+public class JotmComparison extends ProductComparison {
+    public static void main(String[] args) throws RunnerException, CommandLineOptionException, CoreEnvironmentBeanException {
+        JMHConfigJTA.runJTABenchmark(JotmComparison.class.getSimpleName(), args);
+    }
+
+    protected ProductInterface getProductInterface() {
+        return jotm;
+    }
+
+    private ProductInterface jotm = new ProductInterface() {
+        org.objectweb.jotm.Jotm tm;
+
+        @Override
+        public UserTransaction getUserTransaction() throws SystemException {
+            return tm.getUserTransaction();
+        }
+
+        @Override
+        public TransactionManager getTransactionManager() {
+            return tm.getTransactionManager();
+        }
+
+        @Override
+        public String getName() {
+            return "jotm";
+        }
+
+        @Override
+        public String getNameOfMetric() {
+            return outerClassName + "_Jotm";
+        }
+
+        @Override
+        public XAResource getXAResource() {
+            return new XAResourceImpl();
+        }
+
+        @Override
+        public void init() {
+            try {
+                tm = new org.objectweb.jotm.Jotm(true, false);
+            } catch (javax.naming.NamingException e) {
+                throw new RuntimeException(e);
+            }
+        }
+
+        @Override
+        public void fini() {
+            if (tm != null) {
+                org.objectweb.jotm.TimerManager.stop();
+                tm.stop();
+                tm = null;
+            }
+        }
+    };
+}
